+++ src/main/java/org/molgenis/vipannotate/annotation/IntReadValueFunction.java
@@ -4,5 +4,5 @@ import org.molgenis.vipannotate.serialization.MemoryBuffer;
 
 @FunctionalInterface
 public interface IntReadValueFunction {
-  int apply(MemoryBuffer memoryBuffer, int index);
+  long apply(MemoryBuffer memoryBuffer, int index);
 }
