+++ src/main/java/org/molgenis/vipannotate/annotation/OffsetIntAnnotationDecoder.java
@@ -11,13 +11,13 @@ public class OffsetIntAnnotationDecoder implements AnnotationDecoder<IntAnnotati
 
   @Override
   public IntAnnotation decode(MemoryBuffer memBuffer, int annotationIndex) {
-    int value = intReadValueFunction.apply(memBuffer, annotationIndex);
+    long value = intReadValueFunction.apply(memBuffer, annotationIndex);
     return new IntAnnotation(offset + value);
   }
 
   @Override
   public void decodeInto(MemoryBuffer memBuffer, int annotationIndex, IntAnnotation annotation) {
-    int value = intReadValueFunction.apply(memBuffer, annotationIndex);
+    long value = intReadValueFunction.apply(memBuffer, annotationIndex);
     annotation.setValue(offset + value);
   }
 }
